增加管理员管理、辅导员管理、学生管理、成绩管理的异常处理，并在前台页面上使用Alert将其输出。

+++ src/main/java/com/houseyoung/student/service/impl/ScoreServiceImpl.java
@@ -18,33 +18,78 @@ public class ScoreServiceImpl implements ScoreService{
     private ScoreMapper scoreMapper;
 
     @Override
-    public List<ScoreDto> listScore(int classId, String studentId, String courseName){
-        return scoreMapper.listScore(classId, studentId, courseName);
+    public List<ScoreDto> listScore(int classId, String studentId, String courseName) throws Exception{
+        try {
+            return scoreMapper.listScore(classId, studentId, courseName);
+        } catch (Exception e){
+            throw new Exception("系统内部异常");
+        }
     }
 
     @Override
-    public void insert(ScoreDto scoreDto){
-        scoreMapper.insert(scoreDto);
+    public void insert(ScoreDto scoreDto) throws Exception{
+        try {
+            if (scoreDto.getStudentId() == null || scoreDto.getStudentId() == "") {
+                throw new Exception("学号不可为空");
+            }
+            if (scoreDto.getCourseId() == null || scoreDto.getCourseId() == "") {
+                throw new Exception("课程编号不可为空");
+            }
+            //TODO: 判断学号是否存在
+            //TODO: 判断课程编号是否存在
+            scoreMapper.insert(scoreDto);
+        } catch (Exception e){
+            throw e;
+        }
     }
 
     @Override
-    public void delete(int id)  {
-        scoreMapper.delete(id);
+    public void delete(int id) throws Exception{
+        try {
+            if (id <= 0) {
+                throw new Exception("ID不合法");
+            }
+            scoreMapper.delete(id);
+        } catch (Exception e){
+            throw e;
+        }
     }
 
     @Override
-    public ScoreDto queryById(int id) {
-        return scoreMapper.queryById(id);
+    public ScoreDto queryById(int id) throws Exception{
+        try {
+            if (id <= 0) {
+                throw new Exception("ID不合法");
+            }
+            return scoreMapper.queryById(id);
+        } catch (Exception e){
+            throw e;
+        }
     }
 
     @Override
-    public void update(ScoreDto scoreDto){
-        scoreMapper.update(scoreDto);
+    public void update(ScoreDto scoreDto) throws Exception{
+        try {
+            if (scoreDto.getStudentId() == null || scoreDto.getStudentId() == "") {
+                throw new Exception("学号不可为空");
+            }
+            if (scoreDto.getCourseId() == null || scoreDto.getCourseId() == "") {
+                throw new Exception("课程编号不可为空");
+            }
+            //TODO: 判断学号是否存在
+            //TODO: 判断课程编号是否存在
+            scoreMapper.update(scoreDto);
+        } catch (Exception e){
+            throw e;
+        }
     }
 
     @Override
-    public List<ScoreDto> listHimself(String studentId) {
-        return scoreMapper.listHimself(studentId);
+    public List<ScoreDto> listHimself(String studentId) throws Exception{
+        try {
+            return scoreMapper.listHimself(studentId);
+        } catch (Exception e){
+            throw new Exception("系统内部异常");
+        }
     }
-
 }
